Windows, error injector
The main changes in this commit are improvements to the Windows port, which
although it is not fully functional it does now send/receive packets (just not
multiple ones) in both client and server modes.

The error injector has been much improved, it now has nearly fully configurable
behavior with a better random number generator. There are a few improvements to
be made, but this can be used to generate data for testing purposes now.

+++ print.c
@@ -11,7 +11,7 @@ int tftp_packet_process(bool send, uint8_t *buffer, size_t length)
 {
 	assert(buffer);
 	FILE *out = stdout;
-	fprintf(out, "%p(%zu): %s\n", buffer, length, send ? "TX" : "RX");
+	fprintf(out, "%p(%u): %s\n", buffer, (unsigned)length, send ? "TX" : "RX");
 	for(size_t i = 0; i < length; i+= COLUMNS) {
 		for(size_t j = i; j < MIN(length, i+COLUMNS); j++)
 			fprintf(out, "%02x ", buffer[j]);
